+++ engine/src/flutter/lib/web_ui/lib/src/engine/canvaskit/renderer.dart
@@ -363,7 +363,7 @@ class CanvasKitRenderer extends Renderer {
   ui.ParagraphStyle createParagraphStyle({
     ui.TextAlign? textAlign,
     ui.TextDirection? textDirection,
-    int? maxLines,
+    double? maxLines,
     String? fontFamily,
     double? fontSize,
     double? height,
