+++ packages/flutter/lib/src/rendering/paragraph.dart
@@ -329,7 +329,7 @@ class RenderParagraph extends RenderBox
     )
     double textScaleFactor = 1.0,
     TextScaler textScaler = const _UnspecifiedTextScaler(),
-    int? maxLines,
+    double? maxLines,
     Locale? locale,
     StrutStyle? strutStyle,
     TextWidthBasis textWidthBasis = TextWidthBasis.parent,
@@ -643,11 +643,11 @@ class RenderParagraph extends RenderBox
   /// An optional maximum number of lines for the text to span, wrapping if
   /// necessary. If the text exceeds the given number of lines, it will be
   /// truncated according to [overflow] and [softWrap].
-  int? get maxLines => _textPainter.maxLines;
+  double? get maxLines => _textPainter.maxLines;
 
   /// The value may be null. If it is not null, then it must be greater than
   /// zero.
-  set maxLines(int? value) {
+  set maxLines(double? value) {
     assert(value == null || value > 0);
     if (_textPainter.maxLines == value) {
       return;
@@ -1429,7 +1429,7 @@ class RenderParagraph extends RenderBox
       DiagnosticsProperty<TextScaler>('textScaler', textScaler, defaultValue: TextScaler.noScaling),
     );
     properties.add(DiagnosticsProperty<Locale>('locale', locale, defaultValue: null));
-    properties.add(IntProperty('maxLines', maxLines, ifNull: 'unlimited'));
+    properties.add(DoubleProperty('maxLines', maxLines, ifNull: 'unlimited'));
   }
 }
